Remove `undefined` type from state
Consumers of the container should rely on the state always being
defined. With this definitely assigned assertion we break that contract
and instead rely on the derived class to make sure the state is init.
The compiler can't check that for us so this will not prevent runtime
errors, but it is an improvement in the API syntax.